requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
numpy==2.4.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
ultralytics
ctkscrollabledropdownpp
pygrabber